+++ CHANGELOG.md
@@ -19,7 +19,7 @@
   - "cpuLoad": The simulation speed is indicated by a color. If too slow the cpuLoad is shown in yellow.
   - Loading a .p file sets 'topOfStack' automatically. A warning is send if 'topOfStack' is set manually.
   - "zx81UlaOptions" added.
-  - "chroma81" support
+  - "chroma81" support (including colourization file support)
 - "topOfStack": default is now taken from the memory model.
 
 # 3.4.1